Create confUser.py
This was an experiment in getting a loop of random data from a gsheet could you use that to search using your own profile to confuse your search history? this probably would work but might get you banned or worse so probably don't use it the way it's written honestly I haven't even run it was just interesting to write. As I said it was an experiment in the data flow pipeline.